# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Blinker API
  description: Documentation globale de l'API Blinker
  version: 1.0.0
servers:
  - url: "http://localhost:3011/"
  - url: "https://dev.blinker.eterny.fr/"
...
paths:
...
  /users/register:
    post:
      tags:
        - Users
      summary: Créer un nouvel utilisateur
      description: Enregistrer un nouvel utilisateur avec un email, un mot de passe et un nom.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "Password123!"
                name:
                  type: string
                  example: "John Doe"
      responses:
        '201':
          description: Utilisateur créé avec succès.
        '400':
          description: Validation échouée.
        '409':
          description: Conflit, utilisateur existant.
  /users:
    get:
      tags:
        - Users
      summary: Obtenir tous les utilisateurs
      description: Récupère une liste de tous les utilisateurs enregistrés.
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Liste des utilisateurs récupérée avec succès.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Erreur interne du serveur.
    post:
      tags:
        - Users
      summary: Créer un utilisateur
      description: Ajoute un nouvel utilisateur avec un email, un mot de passe sécurisé et un nom.
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Adresse email de l'utilisateur
                password:
                  type: string
                  description: Mot de passe sécurisé (minimum 12 caractères, majuscules, minuscules, chiffres et symboles requis)
                name:
                  type: string
                  description: Nom de l'utilisateur
      responses:
        '201':
          description: Utilisateur créé avec succès.
        '400':
          description: Erreur de validation.
        '409':
          description: Conflit - l'utilisateur existe déjà.
  /users/{id}:
    get:
      tags:
        - Users
      summary: Obtenir un utilisateur par ID
      description: Récupère un utilisateur à partir de son ID unique.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID de l'utilisateur à rechercher.
      responses:
        '200':
          description: Utilisateur trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur interne du serveur.
    put:
      tags:
        - Users
      summary: Mettre à jour un utilisateur
      description: Met à jour les informations d'un utilisateur (nom, email, mot de passe, etc.).
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID de l'utilisateur à mettre à jour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nouveau nom de l'utilisateur
                email:
                  type: string
                  format: email
                  description: Nouvelle adresse email
                password:
                  type: string
                  description: Nouveau mot de passe (au moins 8 caractères)
      responses:
        '204':
          description: Utilisateur mis à jour avec succès.
        '400':
          description: Erreur de validation.
        '404':
          description: Utilisateur non trouvé.
    delete:
      tags:
        - Users
      summary: Supprimer un utilisateur par ID
      description: Supprime un utilisateur en utilisant son ID.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID de l'utilisateur.
      responses:
        '204':
          description: Utilisateur supprimé avec succès.
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur interne du serveur.
  /users/email/{email}:
    get:
      tags:
        - Users
      summary: Obtenir un utilisateur par email
      description: Récupère un utilisateur à partir de son adresse email.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
          description: Email de l'utilisateur à rechercher.
      responses:
        '200':
          description: Utilisateur trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur interne du serveur.
  /users/confirm/{token}:
    get:
      tags:
        - Users
      summary: Confirmer un compte utilisateur
      description: Vérifie un compte utilisateur à l'aide d'un token de confirmation.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: Token de confirmation.
      responses:
        '200':
          description: Compte vérifié avec succès.
        '400':
          description: Lien de confirmation invalide ou expiré.
        '404':
          description: Utilisateur non trouvé.
  /users/send-test-email/:
    post:
      tags:
        - Users
      summary: Envoyer un email de test
      description: Envoie un email de confirmation à l'utilisateur spécifié.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userID:
                  type: string
                  example: "user123"
                subject:
                  type: string
                  example: "Confirmation de votre compte"
                text:
                  type: string
                  example: "Cliquez sur le lien pour confirmer votre compte."
      responses:
        '200':
          description: E-mail de confirmation envoyé avec succès.
        '400':
          description: Erreur de validation.
        '500':
          description: Erreur lors de l'envoi de l'e-mail.
  /users/grant-admin/{userID}:
    post:
      tags:
        - Users
      summary: Accorder le rôle d'administrateur à un utilisateur
      description: Permet à un administrateur de promouvoir un utilisateur au rôle d'administrateur.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userID
          required: true
          schema:
            type: string
            format: uuid
          description: ID de l'utilisateur à promouvoir.
      responses:
        '200':
          description: L'utilisateur a été promu administrateur avec succès.
        '403':
          description: Accès refusé - seul un administrateur peut accorder ce rôle.
        '404':
          description: Utilisateur non trouvé.
  /users/reset-password:
    post:
      tags:
        - Users
      summary: Demander une réinitialisation de mot de passe
      description: Envoie un email de réinitialisation de mot de passe si l'utilisateur existe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "user@example.com"
      responses:
        '200':
          description: Email de réinitialisation envoyé si l'utilisateur existe.
        '400':
          description: Validation échouée.
        '404':
          description: Utilisateur non trouvé.
  /users/reset-password/{token}:
    post:
      tags:
        - Users
      summary: Réinitialiser le mot de passe
      description: Permet à un utilisateur de définir un nouveau mot de passe après avoir demandé une réinitialisation.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
          description: Token de réinitialisation du mot de passe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "SecurePassword123!"
      responses:
        '200':
          description: Mot de passe mis à jour avec succès.
        '400':
          description: Erreur de validation ou token invalide.
        '404':
          description: Utilisateur non trouvé.
  # ---------------------------
  # Endpoints Blinks
  # ---------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'JWT'
      in: 'header'
      name: 'Authorization'
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "user123"
        email:
          type: string
          format: email
          example: "user@example.com"
...